requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bidict==0.22.1
cached-property==1.5.2
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.2.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
openai==1.39.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
python-engineio==4.4.1
python-socketio==5.8.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.7.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
